first round of work to add new biocontainers with aliases
this will preview the addition and test the workflow.
We will likely want to run this once a month (before
the registry is updated) and it cannot be merged until
a new update strategy is in place. I have an idea based
on taking hashes of names, and then breaking those down
equally into chunks to run on different days of the month.
More TBA!

Signed-off-by: vsoch <vsoch@users.noreply.github.com>